chore: move Gleap in its own Hook
