+++ packages/react-dogfood/pages/join/[callId].tsx
@@ -1,14 +1,10 @@
-import { useEffect } from 'react';
-import Gleap from 'gleap';
 import { useRouter } from 'next/router';
 import {
-  Call,
   MediaDevicesProvider,
   StreamVideo,
   useCreateStreamVideoClient,
 } from '@stream-io/video-react-sdk';
 import Head from 'next/head';
-
 import { useCreateStreamChatClient } from '../../hooks';
 import { LoadingScreen, MeetingUI } from '../../components';
 import { getDeviceSettings } from '../../components/DeviceSettingsCaptor';
@@ -16,6 +12,7 @@ import {
   getServerSideCredentialsProps,
   ServerSideCredentialsProps,
 } from '../../lib/getServerSideCredentialsProps';
+import { useGleap } from '../../hooks/useGleap';
 
 const CallRoom = (props: ServerSideCredentialsProps) => {
   const router = useRouter();
@@ -36,59 +33,7 @@ const CallRoom = (props: ServerSideCredentialsProps) => {
     userData: user,
   });
 
-  useEffect(() => {
-    if (gleapApiKey) {
-      Gleap.initialize(gleapApiKey);
-      Gleap.identify(user.name || user.id, {
-        name: user.name,
-      });
-    }
-  }, [gleapApiKey, user.name, user.id]);
-
-  useEffect(() => {
-    if (!gleapApiKey) return;
-
-    Gleap.on('flow-started', () => {
-      try {
-        const { getCurrentValue, ...state } = client.readOnlyStateStore;
-        const data = Object.entries(state).reduce<Record<string, any>>(
-          (acc, [key, observable]) => {
-            if (!!observable && typeof observable.subscribe === 'function') {
-              const value = getCurrentValue<unknown>(observable);
-              if (key === 'activeCall$' && value && value instanceof Call) {
-                // special handling for the active call
-                const call = value;
-                const ignoredKeys = [
-                  // these two are derived from participants$.
-                  // we don't want to send the same data twice.
-                  'localParticipant$',
-                  'remoteParticipants$',
-                ];
-                Object.entries(call.state)
-                  .filter(([k]) => k.endsWith('$') && !ignoredKeys.includes(k))
-                  .forEach(([k, v]) => {
-                    if (!!v && typeof v.subscribe === 'function') {
-                      acc[`${key}.${k}`] = getCurrentValue(v);
-                    } else {
-                      acc[`${key}.${k}`] = v;
-                    }
-                  });
-              } else {
-                acc[key] = value;
-              }
-            }
-            return acc;
-          },
-          {},
-        );
-        console.log('!!State Store', data);
-        Gleap.attachCustomData(data);
-      } catch (e) {
-        console.error(e);
-      }
-    });
-  }, [client.readOnlyStateStore, gleapApiKey]);
-
+  useGleap(gleapApiKey, client, user);
   const deviceSettings = getDeviceSettings();
 
   if (!client) {
